Fixed incorrect png image export
This commit fixes a bug preventing one from exporting PNG images although
specified otherwise in the command line.

The code was relying on `type` instead of `options.format` to detect the
output image's type.

Besides, the quality option is set to undefined when exporting to the PNG
format as this field must not be defined when exporting to PNG.